add loadResourceUtils class
